requirements.txt
pyinquirer==1.0.3
Pillow>=8.2.0